refactor(repository): repository 추상화
